--- tools/packages/wesl/src/WESLTokens.ts
@@ -1,7 +1,0 @@
-/** main set of tokens for WGSL */
-export const mainTokens = {
-  ident: "word",
-  keyword: "keyword",
-  digits: "number",
-  symbol: "symbol",
-};
